requirements.txt
python-telegram-bot==12.7
tensorflow==2.0.0
pillow